implemented searchVenues method.
in order to use the 2dsphere indexing I had to combine latitude and
longitude properties into location property, and refactor all the
references in the project. searchVenues accept lat, lng, radius and limit
as parameters in the query.
I don't know how well the $nearSphere query works, and I could not find a
really good way of testing it. Therefore I only tested if I get more
results with a larger radius.